scheme updt,rm max pwd len,updt del post, fix comment,fix txtb,fix hover

+++ includes/deletePost.php
@@ -18,16 +18,49 @@
    } elseif ($queryPostAutority->rowCount() == 1) {
       $queryPostAutority->closeCursor();
 
-      $queryDeletePost = $connection->prepare("DELETE FROM publicacao WHERE Key_Publicacao=:Chave AND Key_Utilizador=:Utilizador");
-      $queryDeletePost->bindParam(":Chave", $chave, PDO::PARAM_STR);
-      $queryDeletePost->bindParam(":Utilizador", $_SESSION["User_Id"], PDO::PARAM_STR);
-      $queryDeletePost->execute();
+      $queryDeleteLike = $connection->prepare("DELETE FROM gosto WHERE Key_Publicacao=:Chave");
+      $queryDeleteLike->bindParam(":Chave", $chave, PDO::PARAM_STR);
 
-      $queryDeletePost->closeCursor();
-      $connection = null;
+      if ($queryDeleteLike->execute()) {
+         $queryDeleteLike->closeCursor();
 
-      echo "Del";
-      exit();
+         $queryDeleteComments = $connection->prepare("DELETE FROM comentario WHERE Key_Publicacao=:Chave");
+         $queryDeleteComments->bindParam(":Chave", $chave, PDO::PARAM_STR);
+
+         if ($queryDeleteComments->execute()) {
+            $queryDeleteComments->closeCursor();
+
+            $queryDeletePost = $connection->prepare("DELETE FROM publicacao WHERE Key_Publicacao=:Chave AND Key_Utilizador=:Utilizador");
+            $queryDeletePost->bindParam(":Chave", $chave, PDO::PARAM_STR);
+            $queryDeletePost->bindParam(":Utilizador", $_SESSION["User_Id"], PDO::PARAM_STR);
+
+            if ($queryDeletePost->execute()) {
+               $queryDeletePost->closeCursor();
+               $connection = null;
+
+               echo "Del";
+               exit();
+            } else {
+               $queryDeletePost->closeCursor();
+               $connection = null;
+
+               echo "Error";
+               exit();
+            }
+         } else {
+            $queryDeleteComments->closeCursor();
+            $connection = null;
+
+            echo "Error";
+            exit();
+         }
+      } else {
+         $queryDeleteLike->closeCursor();
+         $connection = null;
+
+         echo "Error";
+         exit();
+      }
    } else {
       $queryPostAutority->closeCursor();
       $connection = null;
@@ -35,4 +68,9 @@
       echo "Error";
       exit();
    }
+
+   $connection = null;
+
+   echo "Error";
+   exit();
 ?>
